refactor: 사용하지 않는 Rank 클래스 삭제
