+++ src/test/java/chess/domain/game/state/TurnTest.java
@@ -1,6 +1,6 @@
-package chess.domain.state;
+package chess.domain.game.state;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
